requirements.txt
pydantic
langgraph
langchain
langchain-core
langchain-community
python-dotenv
langchain-groq
arxiv
wikipedia
phidata
streamlit